refactor: move ProjectCard component and add prop types validation
